[13.x] DDF-4659 GSR-4608 Make HandlebarsWfsFeatureTransformer handle XML attributes (#179)
* DDF-4659 Make HandlebarsWfsFeatureTransformer handle XML attributes

HandlebarsWfsFeatureTransformer now adds the XML attributes of all the
elements in the feature to its internal map. The key is of the form
elementName@attributeName to make sure attributes with the same name on
different elements do not overwrite each other. In your mapping JSON you
can specify "elementName@attributeName" as your feature name and use it
in your Handlebars template to get the attribute's value.

* DDF-4659 Use a while loop instead of a for loop