[13.x] GSR-7030 Fix point-radius lat/lon input validation (#243)
Backspacing the entire field results in ''. The Math.max()/Math.min()
operations performed on the value to fit it in the valid latitude or
longitude range return 0 for '' which results in the field being set to
0 even if you erase the whole thing. The current solution in 14.x and
DDF master is to just set the field's value to '' if the final result
after converting to valid is 0.  This doesn't work because you can't set
the field's value to 0 or any value that starts with 0 (i.e. any value
in the range (-1, 1)) because it will immediately be reset to ''.